fix(ci): resolve usage rollup checks
